Allow runtime activation of annunciations.
By default, the program operates in "silent" mode, which allows it to
scrape the website (albeit less often) but not to announce its findings
to GroupMe, nor act upon GroupMe's requests.

Thus, multiple instances of this program can be run by different people,
but with only one in the "active" mode, and the others "dormant" until
the "active" operator changes faction.

Also done:
* Minor runtime-command improvements.
* Text-to-image uses black-on-white instead of white-on-black.
* The GroupMe API key is now obtained from the enviroment.
* Various other minor stuff I forgot.